requirements.txt
PyQt6>=6.6.0
matplotlib>=3.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
ollama>=0.1.7
pytest>=7.0.0
transformers>=4.35.0
torch>=2.0.0
